refactor(filter): reorganize file structure
