+++ packages/admin-ui/src/components/Filter/index.tsx
@@ -1,1 +1,1 @@
-export * from './components/Filter'
+export * from './filter'
